// services/graphql/generated/binding.ts
export interface Query {
  users: <T = Array<User>>(
    args: {
      offset?: Int | null;
      limit?: Int | null;
      where?: UserWhereInput | null;
      orderBy?: UserOrderByInput | null;
    },
    info?: GraphQLResolveInfo | string,
    options?: Options
  ) => Promise<T>;
  user: <T = User>(
    args: { where: UserWhereUniqueInput },
    info?: GraphQLResolveInfo | string,
    options?: Options
  ) => Promise<T>;
}

export interface Mutation {
  createUser: <T = User>(
    args: { data: UserCreateInput },
    info?: GraphQLResolveInfo | string,
    options?: Options
  ) => Promise<T>;
  updateUser: <T = User>(
    args: { data: UserUpdateInput; where: UserWhereUniqueInput },
    info?: GraphQLResolveInfo | string,
    options?: Options
  ) => Promise<T>;
}

export interface Subscription {}

export interface Binding {
  query: Query;
  mutation: Mutation;
  subscription: Subscription;
  request: <T = any>(query: string, variables?: { [key: string]: any }) => Promise<T>;
  delegate(
    operation: 'query' | 'mutation',
    fieldName: string,
    args: {
      [key: string]: any;
    },
    infoOrQuery?: GraphQLResolveInfo | string,
    options?: Options
  ): Promise<any>;
  delegateSubscription(
    fieldName: string,
    args?: {
      [key: string]: any;
    },
    infoOrQuery?: GraphQLResolveInfo | string,
    options?: Options
  ): Promise<AsyncIterator<any>>;
  getAbstractResolvers(filterSchema?: GraphQLSchema | string): IResolvers;
}

export interface BindingConstructor<T> {
  new (...args: any[]): T;
}

export const Binding = makeBindingClass<BindingConstructor<Binding>>({ schema });

/**
 * Types
 */

export type UserOrderByInput =
  | 'createdAt_ASC'
  | 'createdAt_DESC'
  | 'updatedAt_ASC'
  | 'updatedAt_DESC'
  | 'deletedAt_ASC'
  | 'deletedAt_DESC'
  | 'email_ASC'
  | 'email_DESC'
  | 'password_ASC'
  | 'password_DESC'
  | 'roles_ASC'
  | 'roles_DESC'
  | 'firstName_ASC'
  | 'firstName_DESC'
  | 'lastName_ASC'
  | 'lastName_DESC'
  | 'status_ASC'
  | 'status_DESC';

export type UserStatus = 'ACTIVE' | 'INACTIVE';

export interface UserCreateInput {
  email?: String | null;
  password?: String | null;
  roles?: String | null;
  firstName?: String | null;
  lastName?: String | null;
  status: UserStatus;
}

export interface UserUpdateInput {
  email?: String | null;
  password?: String | null;
  roles?: String | null;
  firstName?: String | null;
  lastName?: String | null;
  status?: UserStatus | null;
}

export interface UserWhereInput {
  id_eq?: String | null;
  id_in?: String[] | String | null;
  createdAt_eq?: String | null;
  createdAt_lt?: String | null;
  createdAt_lte?: String | null;
  createdAt_gt?: String | null;
  createdAt_gte?: String | null;
  createdById_eq?: String | null;
  updatedAt_eq?: String | null;
  updatedAt_lt?: String | null;
  updatedAt_lte?: String | null;
  updatedAt_gt?: String | null;
  updatedAt_gte?: String | null;
  updatedById_eq?: String | null;
  deletedAt_all?: Boolean | null;
  deletedAt_eq?: String | null;
  deletedAt_lt?: String | null;
  deletedAt_lte?: String | null;
  deletedAt_gt?: String | null;
  deletedAt_gte?: String | null;
  deletedById_eq?: String | null;
  email_eq?: String | null;
  email_contains?: String | null;
  email_startsWith?: String | null;
  email_endsWith?: String | null;
  email_in?: String[] | String | null;
  password_eq?: String | null;
  password_contains?: String | null;
  password_startsWith?: String | null;
  password_endsWith?: String | null;
  password_in?: String[] | String | null;
  roles_eq?: String | null;
  roles_contains?: String | null;
  roles_startsWith?: String | null;
  roles_endsWith?: String | null;
  roles_in?: String[] | String | null;
  firstName_eq?: String | null;
  firstName_contains?: String | null;
  firstName_startsWith?: String | null;
  firstName_endsWith?: String | null;
  firstName_in?: String[] | String | null;
  lastName_eq?: String | null;
  lastName_contains?: String | null;
  lastName_startsWith?: String | null;
  lastName_endsWith?: String | null;
  lastName_in?: String[] | String | null;
  status_eq?: UserStatus | null;
  status_in?: UserStatus[] | UserStatus | null;
}

export interface UserWhereUniqueInput {
  id?: String | null;
  email?: String | null;
}
// ...
export interface User extends BaseGraphQLObject {
  id: ID_Output;
  createdAt: DateTime;
  createdById: String;
  updatedAt?: DateTime | null;
  updatedById?: String | null;
  deletedAt?: DateTime | null;
  deletedById?: String | null;
  version: Int;
  email: String;
  password?: String | null;
  roles?: String | null;
  firstName?: String | null;
  lastName?: String | null;
  status: UserStatus;
}
// ...
export type Int = number;